ref: auth components added
