+++ src/components/Authentication/Authentication.jsx
@@ -1,5 +1,4 @@
-import { Header, Main } from "./index";
-import { Footer } from "../Footer";
+import { Header, Main, Footer } from "./index";
 
 const Authentication = () => {
     return (
