+++ src/components/Authentication/Register.jsx
@@ -4,26 +4,26 @@ const Register = () => {
             <section className="input-container flex-column m-5">
                 <input id="first-name" className="textbox-content p-5" type="first-name"
                     name="first-name" />
-                <label for="first-name" className="textbox-label m-0">First Name<span
+                <label htmlFor="first-name" className="textbox-label m-0">First Name<span
                         className="required-field">*</span></label>
                 <sub className="p-2 my-2 inactive-check"></sub>
             </section>
             <section className="input-container flex-column m-5">
                 <input id="last-name" className="textbox-content p-5" type="last-name" name="last-name" />
-                <label for="last-name" className="textbox-label m-0">Last Name<span
+                <label htmlFor="last-name" className="textbox-label m-0">Last Name<span
                         className="required-field">*</span></label>
                 <sub className="p-2 my-2 inactive-check"></sub>
             </section>
             <section className="input-container flex-column m-5">
                 <input id="email" className="textbox-content p-5" type="email" name="email" />
-                <label for="email" className="textbox-label m-0">Email<span
+                <label htmlFor="email" className="textbox-label m-0">Email<span
                         className="required-field">*</span></label>
                 <sub className="email-check p-2 my-2 inactive-check"></sub>
             </section>
             <section className="input-container flex-column m-5">
                 <input id="password" className="textbox-content p-5" type="password" name="password" />
                 <i className="fas fa-eye show-password" id="show-password"></i>
-                <label for="password" className="textbox-label m-0">Password<span
+                <label htmlFor="password" className="textbox-label m-0">Password<span
                         className="required-field">*</span></label>
                 <sub className="password-check p-2 my-2 inactive-check"></sub>
             </section>
@@ -32,7 +32,7 @@ const Register = () => {
                     type="password" name="confirm-password" />
                 <i className="fas fa-eye show-password" id="show-password"></i>
 
-                <label for="confirm-password" className="textbox-label m-0">Confirm Password<span
+                <label htmlFor="confirm-password" className="textbox-label m-0">Confirm Password<span
                         className="required-field">*</span></label>
                 <sub className="password-check p-2 my-2 confirm-password-check inactive-check"></sub>
             </section>
