+++ src/components/Authentication/index.js
@@ -1,5 +1,6 @@
 export { Authentication } from "./Authentication";
-export { Header } from "./Header";
+export { Header } from "../Header";
+export { Footer } from "../Footer";
 export { Main } from "./Main";
 export { AuthenticationTabs } from "./AuthenticationTabs";
 export { Login } from "./Login";
